finish up offline test and fix crash bugs
