+++ decompiler/analysis/static_refs.cpp
@@ -35,7 +35,7 @@ bool try_convert_lambda(const Function& parent_function,
             "ignored and is no longer required.",
             lab.name);
       }
-      if (!other_func->ir2.env.has_local_vars()) {
+      if (!other_func->ir2.env.has_local_vars() || !other_func->ir2.top_form) {
         // don't bother if we don't even have vars.
         return false;
       }
