+++ decompiler/config/jak1_ntsc_black_label.jsonc
@@ -4,18 +4,7 @@
   // if you want to filter to only some object names.
   // it will make the decompiler much faster.
   "allowed_objects": [],
-  "banned_objects": [
-    "crates",
-    "puffer",
-    "helix-water",
-    "green-eco-lurker",
-    "seagull",
-    "sunken-pipegame",
-    "snow-ram",
-    "snow-ram-boss",
-    "rolling-race-ring",
-    "part-tester"
-  ],
+  "banned_objects": [],
 
   ////////////////////////////
   // CODE ANALYSIS OPTIONS
