+++ test/offline/offline_test_main.cpp
@@ -1,213 +1,94 @@
-#include <common/link_types.h>
-#include "common/util/FileUtil.h"
-#include "gtest/gtest.h"
+#include <string>
+#include <unordered_set>
+#include <vector>
+#include <memory>
+
+#include "third-party/fmt/format.h"
 #include "common/log/log.h"
-#include "decompiler/Disasm/OpcodeInfo.h"
-#include "decompiler/config.h"
+#include "common/common_types.h"
+#include "common/util/FileUtil.h"
+#include "common/util/json_util.h"
 #include "decompiler/ObjectFile/ObjectFileDB.h"
+#include "common/util/diff.h"
 #include "goalc/compiler/Compiler.h"
 #include "common/util/Timer.h"
-#include <common/util/json_util.h>
 
 namespace fs = std::filesystem;
 
-namespace {
-
-// list of object files to ignore during reference checks
-const std::unordered_set<std::string> g_files_to_skip_compiling = {
-    "timer",            // accessing timer regs
-    "display",          // interrupt handlers
-    "target-snowball",  // screwed up labels, likely cut content
+// command line arguments
+struct OfflineTestArgs {
+  bool dump_current_output = false;
+  std::string iso_data_path;
+  s32 max_files = INT32_MAX;
 };
 
-// the functions we expect the decompiler to skip
-const std::unordered_set<std::string> g_functions_expected_to_reject = {
-    // gcommon
-    "quad-copy!",  // asm mempcy
-    // gkernel
-    "set-to-run-bootstrap",     // kernel context switch
-    "throw",                    // manually sets fp/t9.
-    "throw-dispatch",           // restore context
-    "(method 0 catch-frame)",   // save context
-    "(method 11 cpu-thread)",   // kernel -> user context switch
-    "(method 10 cpu-thread)",   // user -> kernel context switch
-    "reset-and-call",           // kernel -> user
-    "return-from-thread-dead",  // kernel -> user
-    "return-from-thread",       // kernel -> user
-    "return-from-exception",    // ps2 exception -> ps2 user
-    "run-function-in-process",  // temp while stack vars aren't supported.
-    // pskernel
-    "kernel-check-hardwired-addresses",  // ps2 ee kernel debug hook
-    "kernel-read-function",              // ps2 ee kernel debug hook
-    "kernel-write-function",             // ps2 ee kernel debug hook
-    "kernel-copy-function",              // ps2 ee kernel debug hook
-    // math
-    "rand-uint31-gen",  // weird and terrible random generator
-    // bounding-box
-    "(method 9 bounding-box)",   // handwritten asm loop
-    "(method 14 bounding-box)",  // handwritten asm loop
-    // trig
-    "exp", "atan0", "sincos!", "sincos-rad!",
-    // matrix
-    "(method 9 matrix)",  // handwritten asm loop
-    "matrix-axis-sin-cos!", "matrix-axis-sin-cos-vu!",
-    // geometry
-    "circle-circle-xz-intersect",  // unused not bothering
-    // dma-h
-    "dma-count-until-done",  // dma asm loop
-    "dma-sync-with-count", "dma-send-no-scratch", "dma-sync-fast",
-    // dma
-    "symlink2", "symlink3",
-    "dma-sync-hang",  // handwritten asm
-    "vector=",        // asm branching
-    // display
-    "vblank-handler",  // asm
-    "vif1-handler", "vif1-handler-debug",
-    // texture
-    "adgif-shader<-texture-with-update!",  // mips2c
-    // sparticle
-    "sp-launch-particles-var", "particle-adgif", "sp-init-fields!", "memcpy", "sp-process-block-2d",
-    "sp-process-block-3d",
-    // ripple - asm
-    "ripple-execute-init", "ripple-create-wave-table", "ripple-apply-wave-table",
-    "ripple-matrix-scale",
+/*!
+ * Parse command line arguments.
+ */
+OfflineTestArgs parse_args(int argc, char* argv[]) {
+  OfflineTestArgs result;
 
-    // collide-mesh-h
-    "(method 11 collide-mesh-cache)",  // asm
+  for (int i = 1; i < argc; i++) {
+    auto arg = std::string(argv[i]);
+    if (arg == "--dump-mode") {
+      result.dump_current_output = true;
+      continue;
+    }
 
-    // mood
-    "update-mood-lava",       // asm
-    "update-mood-lightning",  // asm
+    if (arg == "--max-files") {
+      i++;
+      if (i >= argc) {
+        fmt::print("--max-files must be followed by an integer\n");
+        exit(1);
+      }
+      result.max_files = atoi(argv[i]);
+      fmt::print("Limiting to {} files\n", result.max_files);
+      continue;
+    }
 
-    // ambient
-    "ambient-inspect",  // asm, weird
+    result.iso_data_path = arg;
+    fmt::print("Using {} for ISO data\n", result.iso_data_path);
+  }
 
-    // background
-    "background-upload-vu0", "draw-node-cull",
+  return result;
+}
 
-    // anim-tester
-    "(method 3 anim-tester)",
-    "anim-tester-save-object-seqs"  // anim-tester -- new basic on the stack
+// json config file data (previously was in source of offline_test_main.cpp)
+struct OfflineTestConfig {
+  std::vector<std::string> dgos;
+  std::unordered_set<std::string> skip_compile_files;
+  std::unordered_set<std::string> skip_compile_functions;
+  std::unordered_map<std::string, std::unordered_set<std::string>> skip_compile_states;
 };
 
-const std::unordered_set<std::string> g_functions_to_skip_compiling = {
-    /// GCOMMON
-    // these functions are not implemented by the compiler in OpenGOAL, but are in GOAL.
-    "abs", "ash", "min", "max", "lognor",
-    // weird PS2 specific debug registers:
-    "breakpoint-range-set!",
-    // inline assembly
-    "valid?",
+/*!
+ * Read and parse the json config file, config.json, located in test/offline
+ */
+OfflineTestConfig parse_config() {
+  auto json_file_path = file_util::get_file_path({"test", "offline", "config.json"});
+  auto json = parse_commented_json(file_util::read_text_file(json_file_path), json_file_path);
+  OfflineTestConfig result;
+  result.dgos = json["dgos"].get<std::vector<std::string>>();
+  result.skip_compile_files = json["skip_compile_files"].get<std::unordered_set<std::string>>();
+  result.skip_compile_functions =
+      json["skip_compile_functions"].get<std::unordered_set<std::string>>();
+  result.skip_compile_states =
+      json["skip_compile_states"]
+          .get<std::unordered_map<std::string, std::unordered_set<std::string>>>();
 
-    /// GKERNEL
-    // asm
-    "(method 10 process)", "(method 14 dead-pool)",
+  if (!result.skip_compile_states.empty()) {
+    fmt::print("skip_compile_states wasn't emtpy. It's not implemented in the decompiler yet\n");
+    exit(1);
+  }
+  return result;
+}
 
-    /// GSTATE
-    "enter-state",  // stack pointer asm
-
-    /// MATH
-    "rand-vu-init", "rand-vu",
-    "rand-vu-nostep",  // random hardware
-
-    // trig
-    "sin-rad",          // fpu acc
-    "cos-rad",          // fpu acc
-    "atan-series-rad",  // fpu acc
-
-    /// VECTOR-H
-    "(method 3 vector)",  // this function appears twice, which confuses the compiler.
-    "vector4-dot",        // fpu acc
-
-    "(method 3 profile-frame)",  // double definition.
-
-    // dma-disasm
-    "disasm-dma-list",  // missing a single cast :(
-
-    // math camera
-    "transform-point-vector!", "transform-point-qword!", "transform-point-vector-scale!",
-
-    // display-h
-    "put-draw-env",
-
-    // geometry
-    "calculate-basis-functions-vector!",  // asm requiring manual rewrite
-    "curve-evaluate!",                    // asm requiring manual rewrite
-    "point-in-triangle-cross",            // logior on floats manual fixup
-
-    // texture
-    "(method 9 texture-page-dir)",         // multiplication on pointers
-    "adgif-shader<-texture-with-update!",  // misrecognized bitfield stuff.
-
-    // asm
-    "invalidate-cache-line",
-
-    // stats-h
-    "(method 11 perf-stat)", "(method 12 perf-stat)",
-
-    // sprite-distorter
-    "sprite-draw-distorters",  // uses clipping flag.
-
-    // sync-info
-    "(method 15 sync-info)",         // needs display stuff first
-    "(method 15 sync-info-eased)",   // needs display stuff first
-    "(method 15 sync-info-paused)",  // needs display stuff first
-
-    // sparticle
-    "lookup-part-group-pointer-by-name",  // address of element in array issue
-
-    // ripple - calls an asm function
-    "ripple-execute",
-
-    "get-task-status",
-
-    // aligner - return-from-thread, currently not supported
-    "(method 9 align-control)",
-
-    // stat collection
-    "start-perf-stat-collection", "end-perf-stat-collection",
-
-    // float to int
-    "(method 10 bsp-header)",
-
-    // multiply defined.
-    "(method 3 sprite-aux-list)",
-
-    // camera
-    "slave-set-rotation!", "v-slrp2!", "v-slrp3!",  // vector-dot involving the stack
-
-    // function returning float with a weird cast.
-    "debug-menu-item-var-make-float",
-
-    // decompiler BUG
-    "level-hint-task-process",
-
-    // anim-tester
-    "(method 3 anim-tester)",
-    "anim-tester-save-object-seqs"  // anim-tester -- new basic on the stack
+struct DecompilerFile {
+  std::filesystem::path path;
+  std::string name_in_dgo;
+  std::string unique_name;
+  std::string reference;
 };
-
-// default location for the data. It can be changed with a command line argument.
-std::string g_iso_data_path = "";
-
-bool g_dump_mode = false;
-
-struct decomp_meta {
-  std::string fileName;
-  std::string fileNameOverride;
-  fs::path filePath;
-};
-
-std::vector<decomp_meta> g_object_files_to_decompile_or_ref_check;
-
-std::vector<std::string> dgos = {"CGO/KERNEL.CGO", "CGO/ENGINE.CGO", "CGO/GAME.CGO", "DGO/BEA.DGO",
-                                 "DGO/INT.DGO",    "DGO/VI1.DGO",    "DGO/VI2.DGO",  "DGO/VI3.DGO",
-                                 "DGO/CIT.DGO",    "DGO/MIS.DGO",    "DGO/JUB.DGO",  "DGO/SUN.DGO",
-                                 "DGO/DEM.DGO",    "DGO/FIN.DGO",    "DGO/JUN.DGO",  "DGO/FIC.DGO",
-                                 "DGO/SNO.DGO",    "DGO/SWA.DGO",    "DGO/MAI.DGO",  "DGO/ROB.DGO",
-                                 "DGO/LAV.DGO",    "DGO/OGR.DGO",    "DGO/TRA.DGO",  "DGO/ROL.DGO"};
-
-}  // namespace
 
 std::string replaceFirstOccurrence(std::string& s,
                                    const std::string& toReplace,
@@ -218,13 +99,10 @@ std::string replaceFirstOccurrence(std::string& s,
   return s.replace(pos, toReplace.length(), replaceWith);
 }
 
-int main(int argc, char** argv) {
-  lg::initialize();
+std::vector<DecompilerFile> find_files(const std::vector<std::string>& dgos) {
+  std::vector<DecompilerFile> result;
 
-  // Determine the files to decompile and reference check by scanning the reference directory
-  // All relevant files are assumed to end with `_REF.g[c|d]`
-  // First rough order them
-  std::vector<decomp_meta> reference_files_rough_order;
+  std::unordered_map<std::string, fs::path> files_with_ref;
   for (auto& p : fs::recursive_directory_iterator(
            file_util::get_file_path({"test", "decompiler", "reference"}))) {
     if (p.is_regular_file()) {
@@ -233,339 +111,156 @@ int main(int argc, char** argv) {
         continue;
       }
       std::string object_name = replaceFirstOccurrence(file_name, "_REF", "");
-      reference_files_rough_order.push_back({object_name, "", p.path()});
+      files_with_ref.insert({object_name, p.path()});
     }
   }
+
+  fmt::print("  Found {} reference files\n", files_with_ref.size());
+
   // use the all_objs.json file to place them in the correct build order
   auto j = parse_commented_json(
       file_util::read_text_file(file_util::get_file_path({"goal_src", "build", "all_objs.json"})),
       "all_objs.json");
+
+  std::unordered_set<std::string> matched_files;
   for (auto& x : j) {
-    auto mapped_name = x[0].get<std::string>();
+    auto unique_name = x[0].get<std::string>();
+
     std::vector<std::string> dgoList = x[3].get<std::vector<std::string>>();
-    for (auto& p : reference_files_rough_order) {
-      if (p.fileName == mapped_name) {
-        // Check to see if we've included atleast one of the DGO/CGOs in our hardcoded list
-        // If not BLOW UP
-        bool dgoValidated = false;
-        for (int i = 0; i < (int)dgoList.size(); i++) {
-          std::string& dgo = dgoList.at(i);
-          // can either be in the DGO or CGO folder, and can either end with .CGO or .DGO
-          if (std::find(dgos.begin(), dgos.end(), fmt::format("DGO/{}.DGO", dgo)) != dgos.end() ||
-              std::find(dgos.begin(), dgos.end(), fmt::format("DGO/{}.CGO", dgo)) != dgos.end() ||
-              std::find(dgos.begin(), dgos.end(), fmt::format("CGO/{}.DGO", dgo)) != dgos.end() ||
-              std::find(dgos.begin(), dgos.end(), fmt::format("CGO/{}.CGO", dgo)) != dgos.end()) {
-            dgoValidated = true;
-          }
+    //    for (auto& p : reference_files_rough_order) {
+    auto it = files_with_ref.find(unique_name);
+    if (it != files_with_ref.end()) {
+      // Check to see if we've included atleast one of the DGO/CGOs in our hardcoded list
+      // If not BLOW UP
+      bool dgoValidated = false;
+      for (int i = 0; i < (int)dgoList.size(); i++) {
+        std::string& dgo = dgoList.at(i);
+        // can either be in the DGO or CGO folder, and can either end with .CGO or .DGO
+        if (std::find(dgos.begin(), dgos.end(), fmt::format("DGO/{}.DGO", dgo)) != dgos.end() ||
+            std::find(dgos.begin(), dgos.end(), fmt::format("DGO/{}.CGO", dgo)) != dgos.end() ||
+            std::find(dgos.begin(), dgos.end(), fmt::format("CGO/{}.DGO", dgo)) != dgos.end() ||
+            std::find(dgos.begin(), dgos.end(), fmt::format("CGO/{}.CGO", dgo)) != dgos.end()) {
+          dgoValidated = true;
         }
-        if (!dgoValidated) {
-          fmt::print(
-              "File [{}] is in the following DGOs [{}], and not one of these is in our list! Add "
-              "it!\n",
-              mapped_name, fmt::join(dgoList, ", "));
-          return 1;
-        }
-        // Hack for working around multi-DGO files
-        if (mapped_name != x[1]) {
-          p.fileNameOverride = x[1];
-        }
-        g_object_files_to_decompile_or_ref_check.push_back(p);
-        break;
+      }
+      if (!dgoValidated) {
+        fmt::print(
+            "File [{}] is in the following DGOs [{}], and not one of these is in our list! Add "
+            "it!\n",
+            unique_name, fmt::join(dgoList, ", "));
+        exit(1);
+      }
+
+      DecompilerFile file;
+      file.path = it->second;
+      file.unique_name = it->first;
+      file.name_in_dgo = x[1];
+      result.push_back(file);
+      matched_files.insert(unique_name);
+    }
+  }
+
+  if (matched_files.size() != files_with_ref.size()) {
+    fmt::print("Error: some REF files were not matched to files in all_objs.json:\n");
+    for (auto& f : files_with_ref) {
+      if (matched_files.count(f.first) == 0) {
+        fmt::print(" {}\n", f.first);
       }
     }
+    exit(1);
   }
 
-  // look for an argument that's not a gtest option
-  bool got_arg = false;
-  int max_files = -1;
-  for (int i = 1; i < argc; i++) {
-    auto arg = std::string(argv[i]);
-    if (arg == "--dump-mode") {
-      g_dump_mode = true;
-      continue;
-    }
-    if (arg == "--max-files") {
-      i++;
-      assert(i < argc);
-      max_files = atoi(argv[i]);
-      printf("Limiting to %d files\n", max_files);
-    }
-    if (arg.length() > 2 && arg[0] == '-' && arg[1] == '-') {
-      continue;
-    }
-    if (got_arg) {
-      printf("You can only specify a single path for ISO data\n");
-      return 1;
-    }
-    g_iso_data_path = arg;
-    lg::warn("Using path {} for iso_data", g_iso_data_path);
-    got_arg = true;
-  }
-
-  if (max_files >= 0) {
-    if ((int)g_object_files_to_decompile_or_ref_check.size() > max_files) {
-      g_object_files_to_decompile_or_ref_check.erase(
-          g_object_files_to_decompile_or_ref_check.begin() + max_files,
-          g_object_files_to_decompile_or_ref_check.end());
-    }
-  }
-
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
+  return result;
 }
 
-class OfflineDecompilation : public ::testing::Test {
- protected:
-  static std::unique_ptr<decompiler::ObjectFileDB> db;
-  static std::unique_ptr<decompiler::Config> config;
-
-  static std::unique_ptr<std::unordered_map<std::string, std::string>> final_output_cache;
-
-  static void SetUpTestCase() {
-    // global setup
-    file_util::init_crc();
-    decompiler::init_opcode_info();
-    config = std::make_unique<decompiler::Config>(decompiler::read_config_file(
-        file_util::get_file_path({"decompiler", "config", "jak1_ntsc_black_label.jsonc"})));
-
-    std::unordered_set<std::string> object_files;
-    for (auto& p : g_object_files_to_decompile_or_ref_check) {
-      std::string fileName = p.fileNameOverride == "" ? p.fileName : p.fileNameOverride;
-      object_files.insert(fileName);
-    }
-    config->allowed_objects = object_files;
-    // don't try to do this because we can't write the file
-    config->generate_symbol_definition_map = false;
-
-    std::vector<std::string> dgo_paths;
-    if (g_iso_data_path.empty()) {
-      for (auto& x : dgos) {
-        dgo_paths.push_back(file_util::get_file_path({"iso_data", x}));
-      }
-    } else {
-      for (auto& x : dgos) {
-        dgo_paths.push_back(file_util::combine_path(g_iso_data_path, x));
-      }
-    }
-
-    db = std::make_unique<decompiler::ObjectFileDB>(dgo_paths, config->obj_file_name_map_file,
-                                                    std::vector<std::string>{},
-                                                    std::vector<std::string>{}, *config);
-
-    // basic processing to find functions/data/disassembly
-    db->process_link_data(*config);
-    db->find_code(*config);
-    db->process_labels();
-
-    // fancy decompilation.
-    db->analyze_functions_ir2({}, *config, {});
-
-    final_output_cache = std::make_unique<std::unordered_map<std::string, std::string>>();
-  }
-
-  static void TearDownTestCase() {
-    db.reset();
-    config.reset();
-    final_output_cache.reset();
-  }
+struct Decompiler {
+  std::unique_ptr<decompiler::ObjectFileDB> db;
+  std::unique_ptr<decompiler::Config> config;
 };
 
-std::unique_ptr<decompiler::ObjectFileDB> OfflineDecompilation::db;
-std::unique_ptr<decompiler::Config> OfflineDecompilation::config;
-std::unique_ptr<std::unordered_map<std::string, std::string>>
-    OfflineDecompilation::final_output_cache;
+Decompiler setup_decompiler(const std::vector<DecompilerFile>& files,
+                            const OfflineTestArgs& args,
+                            const OfflineTestConfig& offline_config) {
+  Decompiler dc;
+  file_util::init_crc();
+  decompiler::init_opcode_info();
+  dc.config = std::make_unique<decompiler::Config>(decompiler::read_config_file(
+      file_util::get_file_path({"decompiler", "config", "jak1_ntsc_black_label.jsonc"})));
 
-/*!
- * Check that the most basic disassembly into files/functions/instructions has succeeded.
- */
-TEST_F(OfflineDecompilation, CheckBasicDecode) {
-  int obj_count = 0;
-  db->for_each_obj([&](decompiler::ObjectFileData& obj) {
-    obj_count++;
-    auto& stats = obj.linked_data.stats;
-    // make sure we decoded all instructions
-    EXPECT_EQ(stats.code_bytes / 4, stats.decoded_ops);
-    // make sure all FP uses are properly recognized
-    EXPECT_EQ(stats.n_fp_reg_use, stats.n_fp_reg_use_resolved);
-  });
-
-  EXPECT_EQ(obj_count, config->allowed_objects.size());
-}
-
-TEST_F(OfflineDecompilation, AsmFunction) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (func.suspected_asm) {
-      if (g_functions_expected_to_reject.find(func.name()) ==
-          g_functions_expected_to_reject.end()) {
-        lg::error("Function {} was marked as asm, but wasn't expected.", func.name());
-        failed_count++;
-      }
-    }
-  });
-  EXPECT_EQ(failed_count, 0);
-}
-
-/*!
- * Test that all functions pass CFG build stage.
- */
-TEST_F(OfflineDecompilation, CfgBuild) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (!func.suspected_asm) {
-      if (!func.cfg || !func.cfg->is_fully_resolved()) {
-        lg::error("Function {} failed cfg", func.name());
-        failed_count++;
-      }
-    }
-  });
-
-  EXPECT_EQ(failed_count, 0);
-}
-
-/*!
- * Test that all functions pass the atomic op construction stage
- */
-TEST_F(OfflineDecompilation, AtomicOp) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (!func.suspected_asm) {
-      if (!func.ir2.atomic_ops || !func.ir2.atomic_ops_succeeded) {
-        lg::error("Function {} failed atomic ops", func.name());
-        failed_count++;
-      }
-    }
-  });
-
-  EXPECT_EQ(failed_count, 0);
-}
-
-/*!
- * Test that all functions pass the type analysis stage
- */
-TEST_F(OfflineDecompilation, TypeAnalysis) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (!func.suspected_asm) {
-      if (!func.ir2.env.has_type_analysis() || !func.ir2.env.types_succeeded) {
-        lg::error("Function {} failed types", func.name());
-        failed_count++;
-      }
-    }
-  });
-
-  EXPECT_EQ(failed_count, 0);
-}
-
-TEST_F(OfflineDecompilation, RegisterUse) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (!func.suspected_asm) {
-      if (!func.ir2.env.has_reg_use()) {
-        lg::error("Function {} failed reg use", func.name());
-        failed_count++;
-      }
-    }
-  });
-
-  EXPECT_EQ(failed_count, 0);
-}
-
-TEST_F(OfflineDecompilation, VariableSSA) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (!func.suspected_asm) {
-      if (!func.ir2.env.has_local_vars()) {
-        lg::error("Function {} failed ssa", func.name());
-        failed_count++;
-      }
-    }
-  });
-
-  EXPECT_EQ(failed_count, 0);
-}
-
-TEST_F(OfflineDecompilation, Structuring) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (!func.suspected_asm) {
-      if (!func.ir2.top_form) {
-        lg::error("Function {} failed structuring", func.name());
-        failed_count++;
-      }
-    }
-  });
-
-  EXPECT_EQ(failed_count, 0);
-}
-
-TEST_F(OfflineDecompilation, Expressions) {
-  int failed_count = 0;
-  db->for_each_function([&](decompiler::Function& func, int, decompiler::ObjectFileData&) {
-    if (!func.suspected_asm) {
-      if (!func.ir2.expressions_succeeded) {
-        lg::error("Function {} failed expressions", func.name());
-        failed_count++;
-      }
-    }
-  });
-
-  EXPECT_EQ(failed_count, 0);
-}
-
-namespace {
-void strip_trailing_newlines(std::string& in) {
-  while (!in.empty() && in.back() == '\n') {
-    in.pop_back();
+  // modify the config
+  std::unordered_set<std::string> object_files;
+  for (auto& file : files) {
+    object_files.insert(file.name_in_dgo);  // todo, make this work with unique_name
   }
-}
-}  // namespace
 
-TEST_F(OfflineDecompilation, Reference) {
-  for (decomp_meta& file : g_object_files_to_decompile_or_ref_check) {
-    std::string fileName = file.fileNameOverride == "" ? file.fileName : file.fileNameOverride;
-    auto& obj_l = db->obj_files_by_name.at(fileName);
-    ASSERT_EQ(obj_l.size(), 1);
+  dc.config->allowed_objects = object_files;
+  // don't try to do this because we can't write the file
+  dc.config->generate_symbol_definition_map = false;
 
-    std::string src = db->ir2_final_out(obj_l.at(0));
-
-    lg::info("Comparing {}...", fileName);
-
-    // NOTE - currently only handles .gc files!
-    auto reference = file_util::read_text_file(file.filePath.string());
-
-    bool can_cache = true;
-    for (auto& func_list : obj_l.at(0).linked_data.functions_by_seg) {
-      for (auto& func : func_list) {
-        if (g_functions_to_skip_compiling.find(func.name()) !=
-            g_functions_to_skip_compiling.end()) {
-          can_cache = false;
-          break;
-        }
-      }
+  std::vector<std::string> dgo_paths;
+  if (args.iso_data_path.empty()) {
+    for (auto& x : offline_config.dgos) {
+      dgo_paths.push_back(file_util::get_file_path({"iso_data", x}));
     }
-
-    if (can_cache) {
-      EXPECT_EQ(final_output_cache->count(fileName), 0);
-      final_output_cache->insert({file.fileName, src});
-    }
-
-    strip_trailing_newlines(reference);
-    strip_trailing_newlines(src);
-
-    if (g_dump_mode) {
-      if (reference != src) {
-        file_util::create_dir_if_needed("./failures");
-        file_util::write_text_file("./failures/" + file.fileName + "_REF.gc", src);
-        EXPECT_TRUE(false);
-      }
-    } else {
-      EXPECT_EQ(reference, src);
+  } else {
+    for (auto& x : offline_config.dgos) {
+      dgo_paths.push_back(file_util::combine_path(args.iso_data_path, x));
     }
   }
+
+  dc.db = std::make_unique<decompiler::ObjectFileDB>(dgo_paths, dc.config->obj_file_name_map_file,
+                                                     std::vector<std::string>{},
+                                                     std::vector<std::string>{}, *dc.config);
+
+  std::unordered_set<std::string> db_files;
+  for (auto& files_by_name : dc.db->obj_files_by_name) {
+    for (auto& f : files_by_name.second) {
+      db_files.insert(f.to_unique_name());
+    }
+  }
+
+  if (db_files.size() != files.size()) {
+    fmt::print("DB file error.\n");
+    for (auto& f : files) {
+      if (!db_files.count(f.unique_name)) {
+        fmt::print("didn't find {}\n", f.unique_name);
+      }
+    }
+    exit(1);
+  }
+
+  return dc;
 }
 
-namespace {
+void disassemble(Decompiler& dc) {
+  dc.db->process_link_data(*dc.config);
+  dc.db->find_code(*dc.config);
+  dc.db->process_labels();
+}
+
+void decompile(Decompiler& dc, const OfflineTestConfig& config) {
+  dc.db->analyze_functions_ir2({}, *dc.config, config.skip_compile_functions);
+}
+
+std::string strip_trailing_newlines(const std::string& in) {
+  std::string out = in;
+  while (!out.empty() && out.back() == '\n') {
+    out.pop_back();
+  }
+  return out;
+}
+
+decompiler::ObjectFileData& get_data(Decompiler& dc,
+                                     const std::string& unique_name,
+                                     const std::string& name_in_dgo) {
+  auto& files = dc.db->obj_files_by_name.at(name_in_dgo);
+  auto it = std::find_if(files.begin(), files.end(), [&](const decompiler::ObjectFileData& data) {
+    return data.to_unique_name() == unique_name;
+  });
+  assert(it != files.end());
+  return *it;
+}
+
 int line_count(const std::string& str) {
   int result = 0;
   for (auto& c : str) {
@@ -575,9 +270,47 @@ int line_count(const std::string& str) {
   }
   return result;
 }
-}  // namespace
 
-TEST_F(OfflineDecompilation, Compile) {
+struct CompareResult {
+  std::vector<std::string> failing_files;
+  int total_files = 0;
+  int ok_files = 0;
+  int total_lines = 0;
+
+  bool total_pass = true;
+};
+
+CompareResult compare(Decompiler& dc, const std::vector<DecompilerFile>& refs, bool dump_mode) {
+  CompareResult compare_result;
+
+  for (const auto& file : refs) {
+    auto& data = get_data(dc, file.unique_name, file.name_in_dgo);
+    std::string result = strip_trailing_newlines(data.full_output);
+    std::string ref = strip_trailing_newlines(file_util::read_text_file(file.path.string()));
+    compare_result.total_files++;
+    compare_result.total_lines += line_count(result);
+    if (result != ref) {
+      compare_result.failing_files.push_back(file.unique_name);
+      compare_result.total_pass = false;
+      fmt::print("Reference test failure on {}:\n", file.unique_name);
+      fmt::print("{}\n", diff_strings(result, ref));
+
+      if (dump_mode) {
+        file_util::create_dir_if_needed("./failures");
+        file_util::write_text_file("./failures/" + file.unique_name + "_REF.gc", result);
+      }
+    } else {
+      compare_result.ok_files++;
+    }
+  }
+
+  return compare_result;
+}
+
+bool compile(Decompiler& dc,
+             const std::vector<DecompilerFile>& refs,
+             const OfflineTestConfig& config) {
+  fmt::print("Setting up compiler...\n");
   Compiler compiler;
 
   compiler.run_front_end_on_file({"decompiler", "config", "all-types.gc"});
@@ -585,29 +318,79 @@ TEST_F(OfflineDecompilation, Compile) {
 
   Timer timer;
   int total_lines = 0;
-  for (decomp_meta& file : g_object_files_to_decompile_or_ref_check) {
-    std::string fileName = file.fileNameOverride == "" ? file.fileName : file.fileNameOverride;
-    if (g_files_to_skip_compiling.find(fileName) != g_files_to_skip_compiling.end()) {
+  for (const auto& file : refs) {
+    if (config.skip_compile_files.count(file.name_in_dgo)) {
+      fmt::print("Skipping {}\n", file.name_in_dgo);
       continue;
     }
 
-    lg::info("Compiling {}...", fileName);
+    fmt::print("Compiling {}...\n", file.unique_name);
 
-    auto& obj_l = db->obj_files_by_name.at(fileName);
-    ASSERT_EQ(obj_l.size(), 1);
+    auto& data = get_data(dc, file.unique_name, file.name_in_dgo);
 
-    const auto& cache = final_output_cache->find(fileName);
-    if (cache != final_output_cache->end()) {
-      const auto& src = cache->second;
+    try {
+      const auto& src = data.output_with_skips;
       total_lines += line_count(src);
       compiler.run_full_compiler_on_string_no_save(src);
-    } else {
-      auto src = db->ir2_final_out(obj_l.at(0), g_functions_to_skip_compiling);
-      total_lines += line_count(src);
-      compiler.run_full_compiler_on_string_no_save(src);
+    } catch (const std::exception& e) {
+      fmt::print("Compiler exception: {}\n", e.what());
+      return false;
     }
   }
   auto time = timer.getSeconds();
-  lg::info("Total Lines Compiled: {}. Lines/second: {:.1f}\n", total_lines,
-           (float)total_lines / time);
+  fmt::print("Total Lines Compiled: {}. Lines/second: {:.1f}\n", total_lines,
+             (float)total_lines / time);
+
+  return true;
+}
+
+int main(int argc, char* argv[]) {
+  fmt::print("Offline Decompiler Test 2\n");
+  lg::initialize();
+
+  fmt::print("Reading config...\n");
+  auto args = parse_args(argc, argv);
+  auto config = parse_config();
+
+  fmt::print("Finding files...\n");
+  auto files = find_files(config.dgos);
+  if (args.max_files < (int)files.size()) {
+    files.erase(files.begin() + args.max_files, files.end());
+  }
+
+  fmt::print("Setting up decompiler and loading files...\n");
+  auto decompiler = setup_decompiler(files, args, config);
+
+  fmt::print("Disassembling files...\n");
+  disassemble(decompiler);
+
+  fmt::print("Decompiling...\n");
+  decompile(decompiler, config);
+
+  fmt::print("Comparing...\n");
+  auto compare_result = compare(decompiler, files, args.dump_current_output);
+  fmt::print("Compared {} lines. {}/{} files passed.\n", compare_result.total_lines,
+             compare_result.ok_files, compare_result.total_files);
+
+  if (!compare_result.failing_files.empty()) {
+    fmt::print("Failing files:\n");
+    for (auto& f : compare_result.failing_files) {
+      fmt::print("  {}\n", f);
+    }
+  }
+
+  bool compile_result = compile(decompiler, files, config);
+
+  if (compare_result.total_pass && compile_result) {
+    fmt::print("Pass!\n");
+    return 0;
+  } else {
+    if (!compile_result) {
+      fmt::print("Compilation failed.\n");
+    }
+    if (!compare_result.total_pass) {
+      fmt::print("Comparison failed.\n");
+    }
+  }
+  return 1;
 }
